backend/requirements.txt
django==4.2.11
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
requests==2.32.3
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
uvicorn==0.30.1
django-filter==23.5